# api/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Venue": {
				"$ref": "#/components/schemas/VenueModel",
				"description": "Model Venue"
			},
...
			"UpdateVenueDto": {
				"$ref": "#/components/schemas/infer_typeofupdateVenueSchema_"
			},
...
			"SafeUser": {
				"$ref": "#/components/schemas/Omit_User.password_"
			},
...
			"UpdateUserDto": {
				"$ref": "#/components/schemas/infer_typeofupdateUserSchema_"
			},
...
			"TournamentRuleDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"tournament_id": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean"
					},
					"points_per_win": {
						"type": "number",
						"format": "double"
					},
					"points_per_draw": {
						"type": "number",
						"format": "double"
					},
					"points_per_loss": {
						"type": "number",
						"format": "double"
					},
					"forfeit_score": {
						"type": "number",
						"format": "double"
					},
					"yellow_cards_suspension": {
						"type": "number",
						"format": "double"
					},
					"max_players_per_team": {
						"type": "number",
						"format": "double"
					},
					"min_players_per_team": {
						"type": "number",
						"format": "double"
					},
					"teams_advance_per_group": {
						"type": "number",
						"format": "double"
					},
					"tiebreaker_order": {
						"items": {
							"$ref": "#/components/schemas/TiebreakerOption"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"user": {
						"properties": {
							"phone": {
								"type": "string",
								"nullable": true
							},
							"email": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"phone",
							"email",
							"name",
							"id"
						],
						"type": "object",
						"nullable": true
					},
					"tournament": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"id",
					"tournament_id",
					"is_active",
					"points_per_win",
					"points_per_draw",
					"points_per_loss",
					"forfeit_score",
					"yellow_cards_suspension",
					"max_players_per_team",
					"min_players_per_team",
					"teams_advance_per_group",
					"tiebreaker_order",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateTournamentRuleDto": {
				"$ref": "#/components/schemas/infer_typeofupdateTournamentRuleSchema_"
			},
...
			"Tournament": {
				"$ref": "#/components/schemas/TournamentModel",
				"description": "Model Tournament"
			},
...
			"UpdateTournamentDto": {
				"$ref": "#/components/schemas/infer_typeofupdateTournamentSchema_"
			},
...
			"Team": {
				"$ref": "#/components/schemas/TeamModel",
				"description": "Model Team"
			},
...
			"SeasonTeam": {
				"$ref": "#/components/schemas/SeasonTeamModel",
				"description": "Model SeasonTeam"
			},
...
			"Season": {
				"$ref": "#/components/schemas/SeasonModel",
				"description": "Model Season"
			},
...
			"UpdateSeasonDto": {
				"$ref": "#/components/schemas/infer_typeofupdateSeasonSchema_"
			},
...
			"Role": {
				"$ref": "#/components/schemas/RoleModel",
				"description": "Model Role"
			},
...
			"UpdateRoleDto": {
				"$ref": "#/components/schemas/infer_typeofupdateRoleSchema_"
			},
...
			"PlayerDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"date_of_birth": {
						"type": "string",
						"format": "date-time"
					},
					"position": {
						"$ref": "#/components/schemas/PlayerPosition"
					},
					"height": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"weight": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"nationality": {
						"type": "string",
						"nullable": true
					},
					"avatar": {
						"type": "string",
						"nullable": true
					},
					"is_active": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"user": {
						"properties": {
							"phone": {
								"type": "string",
								"nullable": true
							},
							"email": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"phone",
							"email",
							"name",
							"id"
						],
						"type": "object",
						"nullable": true
					},
					"user_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"date_of_birth",
					"position",
					"height",
					"weight",
					"nationality",
					"avatar",
					"is_active",
					"created_at",
					"updated_at",
					"user_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreatePlayerDto": {
				"$ref": "#/components/schemas/infer_typeofcreatePlayerSchema_"
			},
...
			"UpdatePlayerDto": {
				"$ref": "#/components/schemas/infer_typeofupdatePlayerSchema_"
			},
...
			"AddPlayerToTeamDto": {
				"$ref": "#/components/schemas/infer_typeofaddPlayerToTeamSchema_"
			},
...
			"BulkDeleteDto": {
				"$ref": "#/components/schemas/infer_typeofbulkDeleteSchema_"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/venues/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Venue"
								}
							}
						}
					}
				},
				"tags": [
					"Venues"
				],
				"security": [
					{
						"jwt": [
							"user",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Venue"
								}
							}
						}
					}
				},
				"tags": [
					"Venues"
				],
				"security": [
					{
						"jwt": [
							"user",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateVenueDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Venues"
				],
				"security": [
					{
						"jwt": [
							"user",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/users/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SafeUser"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SafeUser"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/tournamentrules/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TournamentRuleDto"
								}
							}
						}
					}
				},
				"tags": [
					"TournamentRules"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateTournamentRuleDto"
								}
							}
						}
					}
				},
				"tags": [
					"TournamentRules"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTournamentRuleDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"TournamentRules"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/tournaments/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tournament"
								}
							}
						}
					}
				},
				"tags": [
					"Tournaments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tournament"
								}
							}
						}
					}
				},
				"tags": [
					"Tournaments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTournamentDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Tournaments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/teams/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Team"
								}
							}
						}
					}
				},
				"tags": [
					"Teams"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Team"
								}
							}
						}
					}
				},
				"tags": [
					"Teams"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"coach_name": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"logo": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Teams"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/seasonteams/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SeasonTeam"
								}
							}
						}
					}
				},
				"tags": [
					"SeasonTeams"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"SeasonTeams"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/seasons/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Season"
								}
							}
						}
					}
				},
				"tags": [
					"Seasons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Season"
								}
							}
						}
					}
				},
				"tags": [
					"Seasons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSeasonDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Seasons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/roles/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Role"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Role"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRoleDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/players/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlayerDto"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlayerDto"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePlayerDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/players": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlayerDto"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePlayerDto"
							}
						}
					}
				}
			}
		},
		"/players/{team_id}/team-players": {
			"get": {
				"operationId": "ListTeamPlayers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResult_TeamPlayerDto_"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "team_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "per_page",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "direction",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"in": "query",
						"name": "position",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "approval_status",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "AddPlayerToTeam",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TeamPlayerDto"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "team_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddPlayerToTeamDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "BulkDeleteTeamPlayers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"notFound": {
											"items": {
												"type": "number",
												"format": "double"
											},
											"type": "array"
										},
										"deleted": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"notFound",
										"deleted"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "team_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BulkDeleteDto"
							}
						}
					}
				}
			}
		},
		"/players/{team_id}/team-players/{id}": {
			"get": {
				"operationId": "GetTeamPlayer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TeamPlayerDto"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "team_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateTeamPlayer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TeamPlayerDto"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "team_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTeamPlayerDto"
							}
						}
					}
				}
			}
		},
		"/players/{team_id}/team-players/{id}/approve": {
			"post": {
				"operationId": "ApproveTeamPlayer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TeamPlayerDto"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "team_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/players/{team_id}/team-players/{id}/reject": {
			"post": {
				"operationId": "RejectTeamPlayer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TeamPlayerDto"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "team_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/players/{team_id}/team-players/export": {
			"get": {
				"operationId": "ExportTeamPlayers",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "team_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/players/import-template": {
			"get": {
				"operationId": "DownloadImportTemplate",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}